+++ CHANGES.rst
@@ -74,6 +74,8 @@ Unreleased
 -   The development server uses HTTP/1.1, which enables keep-alive
     connections and chunked streaming responses, when ``threaded`` or
     ``processes`` is enabled. :pr:`2323`
+-   The debugger uses Python's built-in ``traceback`` module in more
+    places. :issue:`1753`
 
 
 Version 2.0.3